Many fixes and refinements
Provisioned items now flash.
Consolidated missile collision determination in missile and turret.
Made mines slightly less commonly dropped by monsters.
Made obstacles slightly less deterrent to monsters, so they'll tend to break things down more often.
Objects are now packaged in a world package.
Humans no longer wander onto items or each other.
Carried humans are no longer damaged by mines.
Fixed mine light color.
Turrets no longer shoot when an enemy is too close.

+++ src/ldActor.cpp
@@ -176,7 +176,10 @@ namespace ld
 			
 			if( m_lightSource )
 			{
-				m_lightSource->destroyWithAnimation();
+				if( !m_lightSource->isDestroying() && !m_lightSource->isMarkedForDeletion() )
+				{
+					m_lightSource->destroyWithAnimation();
+				}
 				m_lightSource = nullptr;
 			}
